+++ Mods ChoGGi/Buildings Always Dusty/metadata.lua
@@ -1,17 +1,17 @@
 return PlaceObj("ModDef", {
 	"title", "Buildings Always Dusty",
+	"id", "ChoGGi_BuildingsAlwaysDusty",
+	"steam_id", "1411885355",
+	"pops_any_uuid", "cfc22b13-e9b3-4b06-8994-35344794cd73",
+	"lua_revision", 1001569,
 	"version", 6,
 	"version_major", 0,
 	"version_minor", 6,
-	"id", "ChoGGi_BuildingsAlwaysDusty",
 	"author", "ChoGGi",
-	"steam_id", "1411885355",
-	"pops_any_uuid", "cfc22b13-e9b3-4b06-8994-35344794cd73",
 	"code", {
 		"Code/Script.lua",
 	},
 	"image", "Preview.png",
-	"lua_revision", 1001569,
 	"has_options", true,
 	"description", [[Buildings will never lose their dust. This is a visual mod, it doesn't affect maintenance.