+++ Mods ChoGGi/Canadian Space Agency 2/metadata.lua
@@ -1,0 +1,33 @@
+return PlaceObj("ModDef", {
+	"title", "Canadian Space Agency 2",
+	"id", "faN8Rlm",
+-- 	"steam_id", "1338969897",
+	"steam_id", "2437837635",
+	"pops_any_uuid", "7aa0de0f-0411-4652-879e-8e061d0a182a",
+	"lua_revision", 1001569,
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "Uncle Urdnot & ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+--~ 	"has_options", true,
+	"TagMissionSponsors", true,
+	"TagColonyLogos", true,
+	"description", [[
+Canadian Space Agency sponsor and logo.
+
+For those looking for a challenging Sponsor.
+More Canadian logos can be found [url=https://steamcommunity.com/sharedfiles/filedetails/?id=1995196689]here[/url].
+
+
+
+
+Original mod made by [url=https://steamcommunity.com/sharedfiles/filedetails/?id=1338969897]Uncle Urdnot[/url].
+I updated it to latest version, as well as:
+Added some vehicle names (suggestions welcome)
+Changed the rocket to the Dragon one (if you have space race).
+]],
+})