+++ Mods ChoGGi/Fix Tourist Overview Softlock/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2434359529",
 	"pops_any_uuid", "2dffb277-a8f8-438b-b9a8-aa78da0333f9",
 	"lua_revision", 1001569,
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
@@ -22,5 +22,7 @@ customSupplyRocket.lua > "OnPressParam", "UIOpenTouristOverview",
 SupplyRocket.lua > SupplyRocket:UIOpenTouristOverview(reward_info)
 instead of getting the reward_info table it gets 1
 
+
+Compatible with Automated Tourism
 ]],
 })