+++ Mods ChoGGi/Meteor Defence Ignores Anomalies/metadata.lua
@@ -1,0 +1,27 @@
+return PlaceObj("ModDef", {
+	"title", "Meteor Defence Ignores Anomalies",
+	"id", "ChoGGi_MeteorDefenceIgnoresAnomalies",
+	"steam_id", "2438415420",
+	"pops_any_uuid", "fb227e70-2835-4b02-a203-56bd61d78c9b",
+	"lua_revision", 1001569,
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagGameplay", true,
+	"description", [[
+MDS Lasers and Defensive Turrets won't shoot down meteors with anomalies and other resources in them.
+Includes mod option to temp disable if one is going to hit a bad spot (needs to be disabled before meteor notifaction).
+
+
+Mod Options: Ignore Anomalies, Ignore Metals, Ignore Polymers
+
+
+Not exactly requested by TheMegaMario.
+]],
+})