+++ Mods ChoGGi/Unlock Research Tech/items.lua
@@ -2,13 +2,13 @@
 
 local PlaceObj = PlaceObj
 local T = T
-local table_concat = table.concat
 
 -- add some descriptions
-local SafeTrans
+local SafeTrans, library_around
 -- use rawget so game doesn't complain about _G
 if rawget(_G, "ChoGGi") then
 	SafeTrans = ChoGGi.ComFuncs.Translate
+	library_around = true
 else
 	local _InternalTranslate = _InternalTranslate
 	local procall = procall
@@ -19,21 +19,33 @@ else
 		procall(function()
 			str = _InternalTranslate(T(varargs))
 		end)
-		return str or T(302535920011424, "Missing text... Nope just needs UICity which isn't around till in-game (ask the devs).")
+		return str or T(302535920011424, "You need to be in-game to see this text (or use my Library mod).")
 	end
 end
 
 local properties = {}
 local c = 0
-
 local TechDef = TechDef
-for id, item in pairs(TechDef) do
-	if item.group ~= "Breakthroughs" and id ~= "None" then
+for id, def in pairs(TechDef) do
+	if def.group ~= "Breakthroughs" and id ~= "None" then
+
+		-- spaces don't work in image tags
+		local image,newline
+		if def.icon:find(" ", 1, true) then
+			image = ""
+			newline = ""
+		else
+			image = "<image " .. def.icon .. ">"
+			newline = "\n\n"
+		end
+
 		c = c + 1
 		properties[c] = PlaceObj("ModItemOptionToggle", {
 			"name", id,
-			"DisplayName", T(item.display_name),
-			"Help", table_concat(T(SafeTrans(item.description, item)) .. "\n\n<image " .. item.icon .. ">"),
+			"DisplayName", SafeTrans(def.display_name) .. (
+				def.icon ~= "UI/Icons/Research/story_bit.tga" and " <right>" .. image or ""
+			),
+			"Help", SafeTrans(def.description, def) .. newline .. image,
 			"DefaultValue", false,
 		})
 	end
@@ -48,15 +60,14 @@ end)
 -- stick res option first
 table.insert(properties, 1, PlaceObj("ModItemOptionToggle", {
 	"name", "TechResearched",
-	"DisplayName", table_concat(T("<yellow>") .. T(302535920011857, "Tech Researched")),
+	"DisplayName", "<yellow>" .. SafeTrans(T(302535920011857, "Tech Researched")),
 	"Help", T(302535920011858, "Turn on to research instead of unlock tech."),
 	"DefaultValue", false,
 }))
 table.insert(properties, 1, PlaceObj("ModItemOptionToggle", {
 	"name", "AlwaysApplyOptions",
-	"DisplayName", T(),
-	"DisplayName", table_concat(T("<yellow>") .. T(302535920011859, "Always Apply Options")),
-	"Help", T(302535920011860, "Unlock/Research techs whenever you load a game/start a new game."),
+	"DisplayName", "<yellow>" .. SafeTrans(T(302535920011859, "Always Apply Options")),
+	"Help", T(302535920011860, "Unlock/Research techs whenever you load a game/start a new game/apply options."),
 	"DefaultValue", false,
 }))